refactor!: move HandErr to submodule of hand module
This makes the code in `lib.rs` a bit smaller and more organized.
